fix: changed volume from double to int to make changing easier

+++ src/main/java/it/unibo/crabinv/Model/audio/SoundService.java
@@ -6,12 +6,12 @@ package it.unibo.crabinv.Model.audio;
 public interface SoundService {
     void playBGM(String musicName);
     void stopBGM();
-    void setBGMVolume(double volume);
+    void setBGMVolume(int volume);
     double getBGMVolume();
     void toggleMuteBGM();
     boolean isBGMMuted();
     void playSfx(String effectName);
-    void setSFXVolume(double volume);
+    void setSFXVolume(int volume);
     double getSFXVolume();
     void toggleMuteSFX();
     boolean isSFXMuted();
